OMAP: DSS2: DSI: Increase HS TX timeout
For some reason screen updates of certain odd widths seem to be triggering
HS TX timeouts on OMAP 3430, even if bigger updates do not. The reason for
this is unknown, but increasing the timeout removes the problem with no
(so far) noticeable problems. I haven't seen this problem on OMAP 3630.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@nokia.com>