Use pgprot_dmacoherent API in OMAPFB
This patch makes use of pgprot_dmacoherent instead of pgprot_noncached API for mmap() functions in OMAPFB

Mapping the shared memory as strongly ordered which also creates an alias
with different mapping attributes. According ARM documentation, if the same memory locations
are marked as having different cachebility attributes, for example by the use of aliases in a virtual
to physical address mapping, behavior is UNPREDICTABLE. L2 cache controller interpretation
for such a mapping is also not defined. This issues will show-up more on SMP system where
the accesses can be in any order.

Signed-off-by: Mayuresh Janorkar <mayur@ti.com>