OMAP: hwmod: Fix state during _setup for module with hardreset
The hwmods with hardreset line and HWMOD_INIT_NO_RESET must be kept
in INITIALIZED state.
They can be properly enabled only if the hardreset line is de-asserted
before.

Fix a small bug that was due to the 'write 1 to reset the status' kind
of behavior.

Check the current reset status to avoid de-asserting the line twice.

Signed-off-by: Benoit Cousson <b-cousson@ti.com>
Cc: Paul Walmsley <paul@pwsan.com>
Cc: Kevin Hilman <khilman@deeprootsystems.com>