DSS2: DSI: disable tearing interrupt before suspend
The DSI external tearing interrupt is now used to invoke
taal_update using an internal workqueue.
This leads to taal suspend taking a long time as taal_update
holds the shared lock.

This patch disables tearing interrupt before the suspend path
tries to access the lock and resumes tearing in  the resume path.

Signed-off-by: Archit Taneja <archit@ti.com>