OMAP4: hwmod: Fix optional clock nodes name in hwmod
Previously, the optional clocks name was used directly
in the hwmod. Unfortunately this clock is the input shared clock
and does not represent the selectable clock in the IP.
Clock framework was updated with virtual clock node that represent the
clock enable in the IP.
Change the hwmod optional clocks name to match these virtual clock nodes.

Signed-off-by: Benoit Cousson <b-cousson@ti.com>
Cc: Paul Walmsley <paul@pwsan.com>
Cc: Kevin Hilman <khilman@deeprootsystems.com>
Cc: Rajendra Nayak <rnayak@ti.com>