OMAP4: hwmod data: Fix missing dss sysconfig & end address
The DSS does not contain any sysconfig register but does contain a
sysstatus register for the overall reset status.

The address range was covering the whole DSS subsystem. Since there
is now a sub-hwmod for IPs inside the DSS, map the top level DSS to
the global sysconfig register only.

Signed-off-by: Benoit Cousson <b-cousson@ti.com>
Cc: Paul Walmsley <paul@pwsan.com>
Cc: Kevin Hilman <khilman@deeprootsystems.com>