usb: musb: Core retention support with musb and in usb boot mode
When the sysboot settings are set I st boot as usb boot, core retention was blocked because
the romcode was enabling the all the possible USBPHY clocks and and USBOTG clocks.
and also the USBPHY was out of sandby mode and phy clock 60Hz was enabled to MUSB.
This was blocking the core retetion even if there is no musb driver loaded.

For device mode:
The PHY is powered down by default in during boot. Phy is powerup when the gadget
driver is loaded. After loading the gadget driver, when there is suspend interrupt enable the SUSPENDM
bit which puts the USBPHY is suspend mode and 60MHz clock is disabled. This allows the core to retention.

Limitations:
With this fix if we connect the cable when the core is in retention, there is omap wakeup.
But the device gets enumerated after the core comes out of the offmode be some other wakeup events.

For HOST mode:

Phy is powered up during boot(this can be done when function driver is loaded). in the idle loop
phy is powereddown before wfi and poweredup back when the core wakesup.

Limitations:
With this fix if we connect the device when the core is in retention, omap is getting woken up and
the device enumeration also fails.

Logged a DVT to fix these limitations.

Signed-off-by: Hema HK <hemahk@ti.com>

Cc: Felipe Balbi <felipe.balbi@nokia.com>
Cc: Tony Lindgren <tony@atomide.com>
Cc: Kevin Hilman <khilman@deeprootsystems.com>