USB: Gadget: For Android, add switch in mass storage
Android (Froyo) has hardcoded a switch on the
volmgr that reads the status (online/offline)
of usb_mass_storage. Refer to: ./system/vold/main.cpp
This patch is to return the state of UMS, through
the sysfs entry:

/sys/devices/virtual/switch/usb_mass_storage/state

as we dont have a fsg->config anymore, we use
common->running, this was set on the
do_set_interface function.

Change-Id:	Ia12dc82cd365aaf3735823f3eef414d3b7f6a805
Signed-off-by: Axel Haslam <axelhaslam@ti.com>