OMAP4: Add a delay sysfs entry for the CMA3000 accelerometer
Add a delay sysfs entry for the accelerometer
Base the measurement mode on the time in ms passed in from
user space.

The ICs measurement modes are
-400Hz
-100Hz
-40Hz

The data rate, if not equal, should always be higher then the value
passed in.  If a data rate of 20ms is passed in then the the data rate which
is higher should be selected.  In this case a value of 100Hz should be selected.

Change-Id: I3360226aa1f55064b46a7623a38b7fc4c6ee78d6
Signed-off-by: Dan Murphy <dmurphy@ti.com>