OMAP: DSS2: Introduce a new "Transition" DSS Display State
DSS2 exposes functions to PM which cut DSS opt clocks if all the panels
are disabled or suspended. Since the time taken for enabling the DSI
engine and panel consumes a lot of time, it is required to change the
state of the DSS panel to a non-disabled/non-suspended state so that
the DSS opt clocks are not touched during the init period.

A new state called OMAP_DSS_DISPLAY_TRANSITION is introduced to
represent this state.