OMAP4: CMA3000 accelerometer delay fix
Fix accelerometer delay implementation at the accel was
being turned off when a value was passed in from the system.

Rearranged the delay structure to the data sheet bit definitions.
Fixed delay finding algo to set the appropriate value.
Use #defines from the header file to set the correct bit in the mode.

Signed-off-by: Dan Murphy <dmurphy@ti.com>