i2c_core: Port i2c PM runtime changes
For devices which are not adapted to runtime PM a call to
pm_runtime_suspended always returns true.

Hence the pm_runtime_suspended checks below prevent legacy
suspend from getting called.

So do a pm_runtime_suspended check only for devices with a
dev_pm_ops populated (which hence do not rely on the legacy
suspend)

Change-Id: I08ae6d0f883713937e2bf53d333877459a842128
Signed-off-by: Rajendra Nayak <rnayak@ti.com>
Signed-off-by: Dan Murphy <dmurphy@ti.com>