CMA3000: Modify power off/on handlers for suspend resume case
The power on handler was turning on the accelerometer regardless
if the accelerometer was on prior to suspending.  It cannot be assumed
that the accelerometers state was on before suspend.  Therefore added
checks in the power off and power on routines to handle if the accelerometer
was on before susspend.

The user space is responsible for turning off the device before suspend can happen.
If the accelerometer is left on while trying to suspend then suspend will fail due
to the user space recieving events from the accelerometer.  Hence why user space suspend
was aborted which meant the kernel suspend thread was never activated to turn off the
accelerometer.

If the accelerometer is properly turned off then the suspend handler will get called
in which case the power off routine will check the state of the accelerometer.  If it was
on then IRQs are disabled and the device is turned off.  If the device was disabled before suspending
then there is nothing to do.

Upon a resume the power on routine was blindly turning on the accelerometer and causing interrupts and
user space failure to re-suspend.  Again it is the user spaces responsibility to turn the device on if
it wants it on.  So when resuming if the device was off before suspend then do not turn the device back on.
Merely return and let the user space take care of re-enabling the device if it wants to.

Change-Id: I26b0a970706ec678b063e4378f4b694c6abe9610
Signed-off-by: Dan Murphy <dmurphy@ti.com>