RCCR and XCCR register values are modified to enable transfer and
receive by default

RCCR and XCCR register values are modified to disable transfer and receive
in the case of "cancelling the transmission". But these values are maintained
even after disabling the clock and enabling the clock. This prevents the
following test cases executed to disable data transfer.

This fix resets the XDISABLE bit and RDSIABLE bit in XCCR and RCCR register
by default.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Shubhrajyoti D <shubhrajyoti@ti.com>