OMAP4: Add early suspend/late resume to touch screen driver
Add the early suspend and late resume handlers and registration
to the synaptics touch screen driver.

The touch driver needs to register for these handlers in Android as the touch screen
needs to be one of the components that go suspends early enough in the suspend
sequence to prevent a false wake up.

Change-Id: Ib9195e056c47c45051511f1ab6108af59e782601
Signed-off-by: Dan Murphy <dmurphy@ti.com>