OMAP4:DSS2: Fix clk debugfs support when DSS clocks are off
Fix the clk debugfs to ignore reporting clock specific information
when the clocks are disabled.

A better fix would be to not print clock information at clock
granularity.  However, for now all clock info is disabled when the
main clock is disabled as at this time the optional clocks are
also disabled.

Change-Id: I7426c6014d6fc64eaef4d0fe2f9fae3bc6291fa8
Signed-off-by: Lajos Molnar <molnar@ti.com>