OMAP/DSS: Make multiple suspends and resume work correctly.
Multiple suspends used to overwrite the activate_after_resume
flags which is incorrect.  Now we ignore subsequent suspends.

Also, disabled state may differ from suspended state (although
for most displays it is the same).  Therefore, call disable on
resume if device is not to be activated after resume.  This is
needed so device state variable corresponds to actual state.

Signed-off-by: Lajos Molnar <molnar@ti.com>