OMAP/DSS2: Added smart enable/disable methods for multi-power displays
Displays are ACTIVE when powered, but may not be fully enabled.
omapdss_display_enable only calls enable for DISABLED devices, so
it cannot be used to fully enable a device that is on low-power
ACTIVE.  These displays can implement a smart_enable method that
is called by omapdss_display_enable regardless of the display state.

These displays should still honor the rule of not changing the
power state on suspended displays (only the resume state).

Also, a smart_is_enabled method should be provided to report the
correct enabled status via sysfs.

Change-Id: I38f7655726faf7a8dc28dd55b48ab735540e6059
Signed-off-by: Lajos Molnar <molnar@ti.com>