OMAP4/DSS: Check for device's manager before calling methods.
When switching between secondary LCD and pico DLP, the devices
manager may be NULL for a short period (or it may be NULL simply
by not setting the manager's display to the new display).

In this situation a NULL reference error may happen.

So we must check if the display has an associated manager before
dereferencing it.

Signed-off-by: Lajos Molnar <molnar@ti.com>