OMAP-DSS: Fix early suspend implementation
Removed the temporary nullifying of the standard suspend and
resume handlers.  Wrapped this in a #ifdef/#ifndef so that the standard
suspend and resume handlers are only used when early suspend is not supported.

Made the early suspend/late resume functions call the standard suspend/resume
implementations as these implemenations were the same.

NOTE: this change adds support for dss suspend/resume if early suspend
handlers are not implemented.  However, the CORE will not hit RET in using
the normal suspend handlers until the DSS/PM fwk issue is resolved.

Change-Id: I7b208f94d05a8fd11a2e12d234a92505eeafed41
Signed-off-by: Lajos Molnar <molnar@ti.com>
Signed-off-by: Dan Murphy <dmurphy@ti.com>