OMAP4: Add early suspend/late resume to bh1780gli driver
Add the early suspend/late resume to the ALS driver.
The ALS driver is not needed until the display is on.  Implementing
this handler will reduce the suspend and resume time when going into retention
as this driver will not be called until the early suspend/late resume is called
by Android

Change-Id: Ieed0e0c7e9de1684bf4033e65a330c8d7e1f6a8c
Signed-off-by: Dan Murphy <dmurphy@ti.com>