Context save and restore of McBSP SYSCONFIG register
When ABE power domain goes to retention, the contents of MCBSP is lost.
The value of SYSCONFIG register in MCBSP is set to SMARTIDLE and ENWAKEUP by
the hwmod framework. But once the ABE power domain goes to retention, the
contents of MCBSP register is lost and set to reset values (NOIDLE and DISABLE
WAKEUP).

This patch creates two API's to save and restore the contents of MCBSP
SYSCONFIG register. Save API is called in probe to save the content of
SYSCONFIG register and restore API is called in request after
pm_runtime_get_sync. Also omap_mcbsp_read() is modified to update the cache
when reading from the register.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>