usb: musb: Enable cable based otg support.
This patch series has below changes.
   Disable the nop transciever call from the panda boardfile
   Enabled the OTG stack by default in omap_4430sdp_defconfig.
   Registered the notification for VBUS and IDGND events
   from TWL6030 OTG transciever.
   Handle the VBUS events(configure the
   control module register to set the session valid,
   VBUS valid and session end bits.
   Configure the control module  IDGND bit to inform mentor controller
   about the change in IDGND.
   Wait for the mentor to be ready as A device or Bdevice based on the IDGND
   input,call the TWL6030 api to set the VBUS on/off.

During port suspend enable the SUSPENDM to suspend the phy  which will cut the
60MHz clock. This will allow the core domain transitioning to retention when
device is supended.

With this changes if the USB device is connected with micro converter
(ID will be grounded)
The the device will be enumerated as hosy.
If the micro cable is connected to host, it is enumerated as device.

Cleaned up the powerdown code as the phy powerdown will be handled based on
the connect/disconnect events.

Signed-off-by: Hema HK <hemahk@ti.com>