staging/ti-st: remove calls to class_unregister
The GPS and FM classes were created using class_create and were never
used in a call to class_register; the call to class_unregister in the
module cleanup function causes a kernel panic when the module is
unloaded.

The call to class_unregister appears to be superfluous so it is removed.

Change-Id: I0eb8102ecbe28f9f73d9d39169e234d738ec2148
Signed-off-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>