DSS2/V4L2: Fix: Do not check omap_dispc_register_isr for errors
Error checking was introduced in streamon and streamoff to return error
from the case where the mainclk is diabled. Doing a CTRL^C in the middle
of streaming returns an error in unregister_isr anyway. So even if the
mainclk is enabled, the irq register and unregister ISRs throw an error.

Hence error checking shouldn't be done in streamon for in streamon and
streamoff

Change-Id: I298ffc8e24883ac2174b41960a91019272290ee8
Signed-off-by: Lajos Molnar <molnar@ti.com>
Signed-off-by: Archit Taneja <archit@ti.com>