OMAP4: SYNTM12 make the IC state configurable in suspend
Forcing the touch IC to be on and scanning during suspend will lead
to excess power consumption during suspend.  The touch screen should
be off during a suspend but there may be cases where the touch IC is desired
to be on for wakeability.

This patch adds a configuration parameter in the touch driver and the board file
to be able to configure the suspend state of the Touch IC.

For Android it is desired that the IC be off.

Change-Id: I414eaa41973ae403c134b277fa0a927fc1c3b5c9
Signed-off-by: Dan Murphy <dmurphy@ti.com>