OMAP: hmwod: Update the sysc_cache in case module context is lost
Do not skip the sysc programming in the hmwod framework based
on the cached  value alone, since at times the module might
have lost context (due to the Powerdomain in which the module
belongs transitions to either Open Switch RET or OFF).

Do not skip the sysc programming in the hmwod framework based on
the cached value alone, since at times the module might have lost
context (due to the Powerdomain in which the module belongs
transitions to either Open Switch RET or OFF).

Signed-off-by: Rajendra Nayak <rnayak@ti.com>
Cc: Paul Walmsley <paul@pwsan.com>
Cc: Benoit Cousson <b-cousson@ti.com>
Cc: Santosh Shilimkar <santosh.shilimkar@ti.com>
Cc: Kevin Hilman <khilman@deeprootsystems.com>
Acked-by: Axel Haslam <axelhaslam@ti.com>