mmc: set constrain when device/host in standard mode
OMAP4 has a constraint where some IPs in some modes have dependency on Core Voltage.
MMC is one of them and MMC/SD/SDIO in Standard mode (24 MHz) and need to chane the to
12 MHz for OPP50%.
This solution is putting the constrain to avoid the low throughput at opp50 and to avoid
the run-time sychronization issue.

Signed-off-by: Sukumar Ghorai <s-ghorai@ti.com>
Signed-off-by: Madhusudhan Chikkature <madhu.cr@ti.com>