USB: MUSB: Hold/Release a wakelock for MUSB on attach/detach
This patch holds a wakelock in musb driver whenever a USB
cable is attached and releases the wakelock on a detach,
thereby not allowing suspend to happen when USB cable is
connected. If a suspend happens with a USB cable connected
it results in un-predicatable behaviors like kernel hang/
crash. This patch resolves the above mentioned issue.

Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>