OMAP4: Clean up checkpatch issues with devices.c
Resolved checkpatch errors with devices.c file.

One checkpatch failure remains but it may be a functional change
include <asm/pmu.h> is not a the prefered header file.
Changing this to include <linux/pmu.h> causes a compilation issue.
This will need to be resolved in a separate commit as it may cause
issues.

Change-Id: I8b5186b9fa269cb5de378196b879797cb30aa856
Signed-off-by: Dan Murphy <dmurphy@ti.com>