SYSLINK: notify - reset notify registration chart correctly
This patch fixes a minor bug in the notify driver unregistration process.
The notify registration chart is an array and stores all the registered
events in ascending order (dictates the notify event priorities). This
array is compacted whenever an event is unregistered. The compaction is
done shifting all the subsequent events to the left. Currently, the next
element to the element being unregistered is not reset correctly, thereby
leading to duplicate registrations in the array. This still leaves a valid
duplicate event when that event is unregistered. This scenario happens
when a lower event is unregistered while a higher event is also present.

Change-Id: I47a04e4d01a5081099117280be286a51afa7d1cf
Signed-off-by: Suman Anna <s-anna@ti.com>