OMAP4: PM: Export control to enable/disable CORE/PER M3 clocks
The CORE and PER M3 post dividers are different from the rest of the
DPLL post dividers as in they go to SCRM, and are used
there to export clocks for instance used by external sensor.

There is no automatic HW dependency in PRCM to manage them. Hence these
two clocks (dpll post dividers) should be managed by SW and explicitly
enabled/disabled.

Add control in clock framework to do so and also remove the existing
handling of these post dividers that is done in cpuidle/suspend path.

Signed-off-by: Rajendra Nayak <rnayak@ti.com>