usb: musb: Internal phy changes under UTMI interface mode check
OMAP supports both UTMI and ULPI modes and internal PHY is used only
in UTMI mode. So All the internal phy specific changes are configured
under the UTMI interface check.
Removed the powering up of the phy from the gadget driver as it was
unnecessary
Moved the powerdown of phy code from gadget driver to platform file.

Defined the offsets and bitfields for the usb control and
dev_conf registers and using them instead of hardcoding.

Signed-off-by: Hema HK <hemahk@ti.com>
Change-Id:  I7894f0d7967cb5d5f4d24dfc65e7aca98bb71f7d