ASoC: McPDM - Use the kernel workqueue to disable Phoenix and McPDM
This is to solve OMAPS00225854. When PDM_DL uses its own queue and has a
delay similar to that of TWL6040 HS ramp down, it is likely that PDM_DL
is deactivated first leading to a pop noise. This is to use kernel
global queue and align with ASoC core and prevent PDM_DL from being
deactivated before TWL6040 HS volume is ramped down.

Change-Id: Ic17e6702384a57f1c684598021dd7e5f3a490b15
Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>