OMAP: PM: Bootup Fixes, Corrections in OPP definitions
With the new clock database, few fixes are needed for bootup.
These fixes are made. Along with this, the OPP definitions are
corrected.  The respective setrate, getrate API's are
corrected according to new OPP definitions. Few clocks in
the 44xx HWMOD database needs to be changed according to new
extended clock database. Also the ipu_pm is using old clk nodes.
These are updated with relevant new clk nodes of clk database.

Change-Id: I41d60afeb79e7b438e468c8fe70e09b8b2155749
Signed-off-by: Avinash.H.M. <avinashhm@ti.com>