OMAP3: Initial Android Defconfig
This is an initial, temporary configuration for the Zoom3.
It is useful while we work on getting 2.6.35 running on the
Zoom3 board, but has many things turned off that we'll need
to turn back on.

For example, CONFIG_PM is turned off to be able to build the
kernel for OMAP3.

Signed-off-by: Steve Schefter <sschefter@ti.com>
Acked-by: Leed Aguilar <leed.aguilar@ti.com>