OMAP3: DSS: Add GPU hwmod definitions, fix DSS clocks, regulator
Along with the hwmod definitions, two temporary work-arounds allow the kernel
to avoid panics and boot to init on OMAP3. The problems are descrbed with
FixMe comments and will have to be resolved when the rest of the OMAP3 DSS
work is completed.

Also, Use vdds_dsi for the regulator name.  It was sometimes vpll2 (the name of
the regulator output) and sometimes vdds_dsi (the OMAP3 input).  Set the
tv_fck, video_fck and dss2_fck clock names and aliases correctly.  Also
needed to have the device as NULL in the omap3xxx_clks[] table since
_add_optional_clock_alias uses NULL for the device (otherwise the clock is
not found, can't be aliased).  Need to enable the dss clocks for OMAP3
during dss_init (previously was just done for OMAP4).

Change-Id: Iccea6435a5f7638713abdcc583c9672fe1c1c114
Signed-off-by: Steve Schefter <sschefter@ti.com>