ASoC: ABE DSP: Add additional check before disabling the event generator
When overrun appears on capture path all the UL ports are closed.
In such condition, the abe_check_activity can return no activity
on ABE side as no ports are active due to underrun. In case of
concurrent end of playback we must protect by checking abe->active
to avoid stop of ABE.

Change-Id: I354e15af9b9528530bc124bd6c8246174ffc6610
Signed-off-by: Sebastien Guiriec <s-guiriec@ti.com>