Revert "SYSLINK : Disable devh notification for tiler events."
This reverts commit 935449094cec4abaf817478cfb88a8be5ef87bb9.

Remove the temporary hack that was added to disable the tiler
event (close) notifications to remote processor.
This is no more needed because of the fix in Tiler driver where
the notification is sent only when there is abnormal termination
of a process without cleaning the resources.

The above issue is fixed by the commit
9ce5953accc4636a9f9c9297197b2f6282690797

Change-Id: I38710367aeb2943f87e4cb4d57633eb166059840
Signed-off-by: Shivananda Hebbar <x0hebbar@ti.com>
Signed-off-by: Hari Kanigeri <h-kanigeri2@ti.com>