OMAP4: PM: Program the CDC TCXO to output CLK1 based on REQ1
The Clock Driver Chip (TCXO) on OMAP4 based SDP needs to
be programmed to output CLK1 based on REQ1 from OMAP.
By default CLK1 is driven based on an internal REQ1INT signal
which is always set to 1.
Doing this helps gate sysclk (from CLK1) to OMAP while OMAP
is in sleep states.
This programming is done onetime at boot.

Change-Id: Ib700f98c4d0d09a848d917c94094570d43c85579
Signed-off-by: Rajendra Nayak <rnayak@ti.com>
Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Signed-off-by: Balaji T K <balajitk@ti.com>