OMAP4 : UART : Fixing the timing response issues with BT chipset
This patch adds the wake-lock to Rx and TX path, to avaoid any
loss of data between the OMAP UART and BT COntroller. In some use
cases, the system idles or suspend immediately after TX data, and
if OFF states are hit, the system BT layer is losing data. Hence
wait for some time before the idle or suspend is hit.

Change-Id: I4a144c18d3fac860519bc0873f1599d72aee8c1a
Signed-off-by: Deepak K <deepak.k@ti.com>