omap: remoteproc: CONFIG flag for self hibernation
A new CONFIG flag is introduce to control IPU self hibernation
the name of the new flag is CONFIG_SYSLINK_IPU_SELF_HIBERNATION
when disable the timers used by IPU to hibernate and
the flag in sh mem to allow hibernation of IPU are not set and IPU
will hibernate only when sending the system suspend.
On the other hand if the flag is set after N seconds of inactivity
and if there is no resource requested by IPU (using ipu_pm) IPU will
hibernate.

Change-Id: I3f14257930a861c8784d7159a821228111b3ba37
Signed-off-by: Miguel Vadillo <vadillo@ti.com>