syslink: ipc: add CONFIG_ flags and traces
Change the Kconfig to separate the CONFIG_ flags for
ipu_pm.

Two CONFIG_ flags are included under IPU_PM:
- CONFIG_SYSLINK_DUCATI_PM
- CONFIG_SYSLINK_IPU_SELF_HIBERNATION
- CONFIG_SYSLINK_IPU_PM_TRACES

Also the Kbuild was modified for the same purpose of
enabling pr_debug traces in ipu_pm module.

IPU self hibernation feature is enabled by default

After 5 seconds (configurable) of IPU in idle it will
send a message to ipu_pm to let it know is ready to be
reset

Change-Id: Ie9d57e7576098f18f638c749599a6264009e85ad
Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Signed-off-by: Leed Aguilar <leed.aguilar@ti.com>