V4L2/videobuf: Fix issue with calling STREAMOFF while not streaming
V4L2 standard does not specify that VIDIOC_STREAMOFF ioctl would
return an error if not streaming.  Instead, it says:

"Return Value

 On success 0 is returned, on error -1 and the errno variable is set
 appropriately:

 EINVAL
    Streaming I/O is not supported, the buffer type is not supported,
    or no buffers have been allocated (memory mapping) or enqueued
    (output) yet."

It can make sense to call STREAMOFF without streaming to reclaim
any queued buffers to the V4L2 device.  E.g. during video playback,
we may queue some buffers to video output device before calling
STREAMON to ensure a steady stream of output frames.  However, if
the video finishes before STREAMON (we have less frames than the
threshold), we would either need to close the device, or use
REQBUFS to request new video buffers.

Using STREAMOFF would be the preferred way to reclaim these buffers,
and there are no other equivalent ways to reclaim them without having
to reopen the device, or remap video buffers.

This is in line with the V4L2 specification:

"The VIDIOC_STREAMOFF ioctl: ... I/O returns to the same state as
 after calling VIDIOC_REQBUFS and can be restarted accordingly."

This patch changes the condition on when to return error in
videobuf_streamoff to whether the queue is empty.  This method is
provided to be used by the streamoff implementations of most
devices, so this change is needed so that the devices can support
STREAMOFF while not streaming.

Change-Id: Ia414e5901884b88a64de3e98bfc633e17918d894
Signed-off-by: Lajos Molnar <molnar@ti.com>