OMAP: HSI: Add IOCTL for SW Reset
This commit provides an IOCTL to reset the HSI block and its
internal DMA Engine.

The old SW reset function was using a schedule_timeout(), which is
not allowed inside a spinlock (from IOCTL).

Furthermore, the schedule_timeout() was made for 10ms, repeatable 20
times, which made it possible to wait 200ms at most. But this is not
needed, as the latency induced by making a HSI register access through
L4 interconnect is much greater than the time it takes to complete the
SW reset, so reading a few times the HSI RESET complete register is enough.

The modified SW reset function doesn't schedule a 10ms delay loop
(iterated max 20 times) to wait for the completion of the SW Reset.

Instead, we simply retry to read the HSI RESET complete, max 5 times.
More than 5 attempts generate an error.
More than 1 attempt generate a warning.

Change-Id: I514a1d6f96d1cdd8c68fea19964f09afb5e81293
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>