OMAP3: DSS: resolve earlier work-arounds
The initial OMAP 3 patch included a couple of FixMes.  This checking fixes
those.  A different order is required for initialization (dispc first).  Also,
the structure type of the data poited to by platform_data is not consistent
for all devices.  The correct one is used and later use of pdata comes from
core.pdata rather than the varying type pdev->dev.platform_data.

Change-Id: Ie17c30959782f584ec697fd9876ef6bbbb44d0f7
Signed-off-by: Steve Schefter <sschefter@ti.com>