OMAP4: DSS: Framebuffer: UI cloning to HDMI
This patch uses one video pipeline to render UI frmaebuffer on HDMI
along with LCD through gfx. By default, there would be no scaling of
the framebufffer so UI would be only shown in the native size on the
HDMI however, a sysfs entry has been created to allow the user to fit
to the screen so that the UI covers the full HDMI screen.

Change-Id: I5b33015ddf66e06b28ae505ba9478c3c9bd4d234
Signed-off-by: Sujeet Kumar Baranwal <s-baranwal@ti.com>