OMAP4: HSI: Fixed invalid resource for max DMA channels available
Porting HSI driver to omap_device/HWMOD has changed the way HSI
resources are defined and used.

Previously there was a static struct resource hsi_resources[] {}
containing a field which defined the max DMA channels available
for the device (SSI or HSI).

This value was defined using the IORESOURCE_DMA flag in an indirect
way. When porting HSI driver to HWMOD, this field was no more present
in HWMOD table, producing an error at boot time : "HSI device has no
GDD channel count resource (use 8 as default)"

Additionaly, now that we have moved to omap_device/HWMOD, every device
shall use the HWMOD the same way, and it is not clean (I think) to use
the HWMOD to pass platform_data to the driver.

So I have extended the hsi_platform_data structure to include a new
field containing the number of DMA channels availables.

This value should (normally) be 8 for SSI and 16 for HSI.
Default value when undefined is 8.

Change-Id: I611ebfdd410f23b23ffef3a018873d06bcf198e6
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>