usb: musb: Removed the omap_write/readl usage in the musb driver
Removed the usage of omap_readl/writel in the musb driver and using
the __raw_real/writel with ioremapped addresses for writing to the
control registers and musb registers in the platform layer.

Also defined a function to enable/disable the phy clock and removed
the hardcoding of CLK_ctrl register for phy clock enable/disable.

Change-Id: Iae26b710b56ae83e59d9a02e8d3aabe85413b2b0
Signed-off-by: Hema HK <hemahk@ti.com>