usb: musb: supporting core retention when host suspends the device
OMAP musb as client, when host suspends the device, the core domain
transition should happen as there will not be any transfers.
As a host, when device is supended the core domain transition should be
allowed.
This patch add the support to allow the core retention when the connected
device is suspended or host suspends the musb.

Tested:
Tested device mode when the host suspends the musb, with global suspend
resume, core is transition to retention and when system is resumed,
the resume of usb device is successful.

Host mode: When devices are suspended core is getting transitioned to
retention.Remote wakeup is working only when the omap is wokeup from
retention.
If the omap is in retention, USB remote wakeup is not able to wake the omap.
Need to check with hardware team whether the sWakeup works in retention or not.

This patch is tested with removing the wake_lock/unlock in the musb driver
in connect/disconnect events.

Change-Id: I8194334e5f8ebc2850a9d3136752f8f3d80f83bc
Signed-off-by: Hema HK <hemahk@ti.com>