usb: musb: USB remote wakeup support using
iopad in host mode

When OMAP is retention, and USB device wants to wakeup the system,
usb remote wakeup should be used to wakeup the omap and resume the
usb device. This patch adds the support to wakeup the omap usinf DP
like as iopad wakeup.

Testing: suspend the OMAP using global suspend, if the usb device
connected is suspendable then the device is suspended and core is
in retention.Now press the key or click the mouse button which
will wakeup the OMAP using IOPAD and usb device gets resumed.

Change-Id: Ifc0d0fe8e8dedf253f1799664775a36499479b22
Signed-off-by: Hema HK <hemahk@ti.com>