usb: musb: Fix for musb when core transitions to off and wakeup
After the coreo-ff musb was not functional because few registers
context save/restore was missing.
Added the context save/restore of the all the required registers to
make musb functional after wakeup.
Holding wake-lock in the device connect as there is spurious interrupts
if core-off with global suspend is allowed with device connected.
So now with this implementation when the device is connected there will
not be global suspend/resume kicked in.
For idle path, when the device is connected, core can still transition
upto OSWR. holding a c state constrint for offmode.

[Leed] Fix checkpatch errors, file change mode (644->755)

Change-Id: I14ac867c00d4c3f24e9223f5068e3d9e23891e95
Signed-off-by: Hema HK <hemahk@ti.com>
Acked-by: Leed Aguilar <leed.aguilar@ti.com>