OMAP: dmtimer: optimize omap2_dm_timer_set_src
There is no need to maintain struct clk {} because they can be
dynamically obtained using clk_get() with appropriate params.
We also get rid of clock names array in mach-omap2 because
with proper alias naming in clock44xx_data.c we can directly
use the names in the function in generic way common to all
platfoms. With this we also get rid of omap2_dm_timer_setup
function which used to populate the struct clk{}.

Modified the timer index to start from 1 instead of 0.
Changed the name from dmtimer to omap-timer.

Cleaned-up the probe() and remove() functions.

Change-Id: I785574a59af5369e93b870f9ca47eaf1565c74be
Signed-off-by: Tarun Kanti DebBarma <tarun.kanti@ti.com>